feat: extarcting envgene-shared module
