+++ devtools/tests/build_modules.sh
@@ -27,3 +27,4 @@ fi
 install_and_clean "envgene" "envgenehelper"
 install_and_clean "artifact-searcher" "artifact_searcher"
 install_and_clean "external-cred-provision" "external_cred_provision"
+install_and_clean "envgene-shared" "envgene_shared"
