+++ modules/envgene-shared/envgene_shared/tests/crypto/test_crypt.py
@@ -1,16 +1,16 @@
 import os
 import copy
+from pathlib import Path
 import subprocess
 import pytest
 from subprocess import SubprocessError
 
 from ruyaml import CommentedMap
 
-from .collections_helper import compare_dicts
-
-from .crypt import decrypt_file, encrypt_file, is_encrypted, _parallel_cred_op
-from .file_helper import check_file_exists, writeToFile
-from .yaml_helper import openYaml, set_nested_yaml_attribute, writeYamlToFile
+from envgene_shared.utils.collections_utils import compare_dicts
+from envgene_shared.crypto.crypt import decrypt_file, encrypt_file, is_encrypted, _parallel_cred_op, detect_crypt_backend_from_file
+from envgene_shared.utils.file_utils import check_file_exists, writeToFile
+from envgene_shared.utils.yaml_utils import openYaml, writeYamlToFile
 
 TEST_CONTENT = """\
 first_cred:
@@ -26,6 +26,7 @@ second_cred:
 TEST_FILE = 'test_data/test_crypt.yaml'
 TEST_FILE_OLD = 'test_data/old_test_crypt.yaml'
 NOT_EXISTING_TEST_FILE = 'test_data/not_existing_test_crypt.yaml'
+EFFECTIVE_SET_FILE = 'test_data/effective-set/runtimee/credentials.yaml'
 
 crypt_test_data = [
     {
@@ -46,6 +47,11 @@ crypt_test_data = [
 ]
 crypt_functions_data = [ decrypt_file, encrypt_file ]
 
+@pytest.fixture(autouse=True)
+def schema_dir(monkeypatch):
+    schema_dir = Path(__file__).resolve().parents[2] / "schemas"
+    monkeypatch.setenv("JSON_SCHEMAS_DIR", str(schema_dir))
+
 def reset_test_files():
     writeToFile(TEST_FILE, TEST_CONTENT)
     if os.path.exists(NOT_EXISTING_TEST_FILE):
@@ -60,6 +66,16 @@ def crypt_kwargs(request):
     crypt_kwargs.update(request.param)
 
     yield crypt_kwargs
+
+
+@pytest.fixture
+def effective_set_setup():
+    effective_dir = os.path.dirname(EFFECTIVE_SET_FILE)
+    os.makedirs(effective_dir, exist_ok=True)    
+    yield    
+    if os.path.exists(EFFECTIVE_SET_FILE):
+        os.remove(EFFECTIVE_SET_FILE)
+
 
 def test_basic(crypt_kwargs):
     init_yaml, enc_yaml, dec_yaml = run_encryption_cycle(crypt_kwargs)
@@ -114,8 +130,8 @@ def test_ignore_crypt(crypt_kwargs):
     new_yaml = encrypt_file(**crypt_kwargs, ignore_is_crypt=True, is_crypt=False)
     assert init_yaml != new_yaml
 
-    new_yaml = decrypt_file(**crypt_kwargs, ignore_is_crypt=False, is_crypt=False)
-    assert init_yaml != new_yaml # should not try to decrypt with is_crypt false, so new_yaml is still encrypted
+    with pytest.raises(ValueError, match="Parameter crypt is set to false"):
+        decrypt_file(**crypt_kwargs, ignore_is_crypt=False, is_crypt=False)
     new_yaml = decrypt_file(**crypt_kwargs, ignore_is_crypt=True, is_crypt=False)
     assert init_yaml == new_yaml
 
@@ -186,7 +202,8 @@ def test_minimize_diff(crypt_kwargs):
 
     # test with one change
     new_content = copy.deepcopy(initial_content)
-    set_nested_yaml_attribute(new_content, 'first_cred.data.secret', 'new-value')
+    # set_nested_yaml_attribute(new_content, 'first_cred.data.secret', 'new-value')
+    new_content["first_cred"]["data"]["secret"] = "new-value"
     writeYamlToFile(cred_file, new_content)
     new_enc_content = encrypt_file(**crypt_kwargs, minimize_diff=True, old_file_path=old_cred_file)
 
@@ -201,3 +218,65 @@ def test_minimize_diff(crypt_kwargs):
     # test wrong parameter combination
     with pytest.raises(ValueError):
         encrypt_file(**crypt_kwargs, minimize_diff=True)
+
+
+def test_encrypt_file_scenarios(crypt_kwargs, effective_set_setup):
+
+    backend = crypt_kwargs.get('crypt_backend')
+    if not backend:
+        return
+
+    local_kwargs = crypt_kwargs.copy()
+    encrypt_file(**local_kwargs)
+    
+    standard_yaml = openYaml(TEST_FILE)
+    assert standard_yaml["first_cred"]["type"] == "secret"
+    assert standard_yaml["first_cred"]["data"]["secret"] != "token-placeholder-123"
+
+    mixed_content = (
+        "api_key: plain-literal-text-token\n"
+        "token: ref+aws/secret/token\n"
+        "ESO_TOKEN:\n"
+        "  normalizedSecretName: test_cluster_01/env-1/dummy-token\n"
+        "  secretStoreId: default-store\n"
+    )
+    writeToFile(EFFECTIVE_SET_FILE, mixed_content)
+    
+    local_kwargs['file_path'] = EFFECTIVE_SET_FILE
+    encrypt_file(**local_kwargs)
+    
+    mixed_yaml = openYaml(EFFECTIVE_SET_FILE)
+
+    assert mixed_yaml['token'] == "ref+aws/secret/token"
+    assert mixed_yaml['ESO_TOKEN']['secretStoreId'] == "default-store"
+    assert mixed_yaml['ESO_TOKEN']['normalizedSecretName'] == "test_cluster_01/env-1/dummy-token"
+    assert mixed_yaml['api_key'] != "plain-literal-text-token"
+
+    pure_literal_content = (
+        "api_key: plain-literal-text-token\n"
+        "db_password: plain-password-123\n"
+    )
+    writeToFile(EFFECTIVE_SET_FILE, pure_literal_content)
+    encrypt_file(**local_kwargs)
+    
+    full_yaml = openYaml(EFFECTIVE_SET_FILE)
+    assert full_yaml['api_key'] != "plain-literal-text-token"
+    assert full_yaml['db_password'] != "plain-password-123"
+
+
+def test_decrypt_auto_detects_backend_from_file(crypt_kwargs):
+    backend = crypt_kwargs.get('crypt_backend')
+    if not backend:
+        return
+
+    encrypt_file(**crypt_kwargs)
+    assert is_encrypted(TEST_FILE, backend)
+
+    wrong_backend = "Fernet" if backend == "SOPS" else "SOPS"
+    local_kwargs = crypt_kwargs.copy()
+    local_kwargs['crypt_backend'] = wrong_backend
+    
+    decrypt_file(**local_kwargs)
+    
+    decrypted_yaml = openYaml(TEST_FILE)
+    assert decrypted_yaml["first_cred"]["data"]["secret"] == "token-placeholder-123"
