+++ modules/envgene-shared/pyproject.toml
@@ -1,0 +1,53 @@
+[build-system]
+requires = ["setuptools>=61"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "envgene-shared"
+version = "0.0.1"
+description = "Shared credential discovery, encryption, and decryption for EnvGene"
+requires-python = "~=3.12"
+classifiers = [
+    "Programming Language :: Python :: 3"
+]
+dependencies = [
+    "cryptography==46.0.5",
+    "jsonschema==4.24.1",
+    "ruyaml==0.91.0",
+]
+
+[project.optional-dependencies]
+dev = [
+    "pytest==8.3.5",
+    "ruff==0.11.0",
+]
+
+[tool.setuptools.packages.find]
+where = ["."]
+
+[tool.setuptools.package-data]
+envgene_shared = [
+    "schemas/*.json",
+    "py.typed"
+]
+
+[tool.black]
+line-length = 120
+skip-string-normalization = true
+
+[tool.mypy]
+warn_return_any = true
+warn_unused_configs = true
+strict_optional = true
+strict_equality = true
+warn_no_return = true
+check_untyped_defs = true
+warn_redundant_casts = true
+show_error_codes = true
+implicit_reexport = false
+warn_unreachable = true
+disallow_incomplete_defs = true
+disallow_any_generics = true
+disallow_untyped_defs = true
+enable_error_code = "truthy-function"
+warn_unused_ignores = true
