+++ modules/envgene/envgenehelper/crypt.py
@@ -1,235 +1,5 @@
-import os
-import re
-import time
-from concurrent.futures import ThreadPoolExecutor, as_completed
-from os import getenv, path
-from typing import Callable
-
-from .config_helper import get_envgene_config_yaml
-from .yaml_helper import openYaml, get_empty_yaml
-from .file_helper import check_file_exists, get_files_with_filter
-from .logger import logger
-from .collections_helper import split_multi_value_param
-
-from .crypt_backends.fernet_handler import crypt_Fernet, extract_value_Fernet, is_encrypted_Fernet
-from .crypt_backends.sops_handler import crypt_SOPS, extract_value_SOPS, is_encrypted_SOPS
-
-BASE_DIR = getenv('CI_PROJECT_DIR', os.getcwd())
-VALID_EXTENSIONS = re.compile(r'\.ya?ml$')
-TARGET_REGEX = re.compile(r'(^credentials$|creds$)')
-TARGET_DIR_REGEX = re.compile(r'/[Cc]redentials(/|$)')
-TARGET_PARENT_DIRS = re.compile(r'/(configuration|environments)(/|$)')
-
-CRYPT_FUNCTIONS = {
-    'SOPS': crypt_SOPS,
-    'Fernet': crypt_Fernet
-}
-
-IS_ENCRYPTED_FUNCTIONS = {
-    'SOPS': is_encrypted_SOPS,
-    'Fernet': is_encrypted_Fernet,
-}
-
-EXTRACT_FUNCTIONS = {
-    'SOPS': extract_value_SOPS,
-    'Fernet': extract_value_Fernet
-}
-
-
-def get_configured_encryption_type():
-    return get_crypt_backend(), get_crypt()
-
-
-def _handle_missing_file(file_path, default_yaml, allow_default):
-    if check_file_exists(file_path):
-        return 0  # sentinel value
-    if not allow_default:
-        raise FileNotFoundError(f"{file_path} not found or is not a file")
-    return default_yaml()
-
-
-def decrypt_file(file_path, *, secret_key=None, in_place=True, public_key=None, crypt_backend=None,
-                 ignore_is_crypt=False,
-                 default_yaml: Callable = get_empty_yaml, allow_default=False, is_crypt=None,
-                 load_result=True, **kwargs):
-    res = _handle_missing_file(file_path, default_yaml, allow_default)
-    if res != 0:
-        return res
-    is_crypt = is_crypt if is_crypt is not None else get_crypt()
-    if not ignore_is_crypt and not is_crypt:
-        logger.info("'crypt' is set to 'false', skipping decryption")
-        if load_result:
-            return openYaml(file_path)
-        return None
-    crypt_backend = crypt_backend if crypt_backend is not None else get_crypt_backend()
-    return CRYPT_FUNCTIONS[crypt_backend](
-        file_path=file_path, secret_key=secret_key, in_place=in_place,
-        public_key=public_key, mode='decrypt', load_result=load_result, **kwargs
-    )
-
-
-def encrypt_file(file_path, *, secret_key=None, in_place=True, public_key=None, crypt_backend=None,
-                 ignore_is_crypt=False, is_crypt=None,
-                 minimize_diff=False, old_file_path=None, default_yaml: Callable = get_empty_yaml, allow_default=False,
-                 load_result=True, **kwargs):
-    if minimize_diff:
-        if not old_file_path:
-            raise ValueError('minimize_diff was set to true but old_file_path was not specified')
-        if not check_file_exists(old_file_path):
-            minimize_diff = False
-            logger.warning(f"Cred file at {old_file_path} doesn't exist, minimize_diff parameter is ignored")
-        elif not is_encrypted(old_file_path, crypt_backend):
-            minimize_diff = False
-            logger.warning(f"Cred file at {old_file_path} is not encrypted, minimize_diff parameter is ignored")
-    res = _handle_missing_file(file_path, default_yaml, allow_default)
-    if res != 0:
-        return res
-    is_crypt = is_crypt if is_crypt is not None else get_crypt()
-    if not ignore_is_crypt and not is_crypt:
-        logger.info("'crypt' is set to 'false', skipping encryption")
-        if load_result:
-            return openYaml(file_path)
-        return None
-    crypt_backend = crypt_backend if crypt_backend is not None else get_crypt_backend()
-    return CRYPT_FUNCTIONS[crypt_backend](
-        file_path=file_path, secret_key=secret_key, in_place=in_place,
-        public_key=public_key, mode='encrypt', minimize_diff=minimize_diff,
-        old_file_path=old_file_path, load_result=load_result, **kwargs
-    )
-
-
-def extract_encrypted_data(file_path, attribute_str):
-    """
-    @param file_path: path to a file
-    @param attribute_str: dot separated path to an attribute 'path.to.an.attribute'
-    @return: decrypted value
-    """
-    crypt_backend = get_crypt_backend()
-    return EXTRACT_FUNCTIONS[crypt_backend](file_path, attribute_str)
-
-
-def is_cred_file(fp: str) -> bool:
-    name = os.path.basename(fp)
-    name_without_ext = os.path.splitext(name)[0]
-    parent_dirs = os.path.dirname(fp)
-    if not VALID_EXTENSIONS.search(name):
-        return False
-    if not TARGET_PARENT_DIRS.search(parent_dirs):
-        return False
-    if TARGET_REGEX.search(name_without_ext) or re.search(TARGET_DIR_REGEX, parent_dirs):
-        return True
-    return False
-
-
-def get_all_necessary_cred_files() -> set[str]:
-    env_names = getenv("ENV_NAMES", None)
-    if not env_names:
-        logger.info("ENV_NAMES not set, running in test mode")
-        return get_files_with_filter(BASE_DIR, is_cred_file)
-    env_names_list = split_multi_value_param(env_names)
-
-    sources = set()
-    sources.add("configuration")
-    sources.add(path.join("environments", "credentials"))
-
-    for env_name in env_names_list:
-        cluster, env = env_name.strip().split("/")
-        env_specific_source_locations = ["credentials", "cloud-passport", "cloud-passports",
-                                         env]  # relative to BASE_DIR/<cluster_name>/
-        for location in env_specific_source_locations:
-            sources.add(path.join("environments", cluster, location))
-
-    cred_files = set()
-    for source in sources:
-        source = path.join(BASE_DIR, source)
-        if not path.exists(source):
-            continue
-        cred_files.update(get_files_with_filter(source, is_cred_file))
-
-    return cred_files
-
-
-def is_encrypted(file_path, crypt_backend=None):
-    CRYPT_BACKEND = get_crypt_backend()
-    crypt_backend = crypt_backend if crypt_backend else CRYPT_BACKEND
-    return IS_ENCRYPTED_FUNCTIONS[crypt_backend](file_path)
-
-
-def check_for_encrypted_files(files):
-    err_msg = "Parameter crypt is set to false in config, but this cred file is encrypted: {}"
-    for f in files:
-        if is_encrypted(f):
-            raise ValueError(err_msg.format(f))
-
-
-def _batch_cred_op(files, op_func, **kwargs):
-    backend = kwargs.get('crypt_backend') or get_crypt_backend()
-    if backend == 'Fernet':
-        for file_path in sorted(files):
-            op_func(file_path, **{**kwargs, 'load_result': False})
-        return
-    _parallel_cred_op(files, op_func, **kwargs)
-
-
-def _parallel_cred_op(files, op_func, **kwargs):
-    file_list = sorted(files)
-    if not file_list:
-        return
-    if kwargs.get('minimize_diff'):
-        max_workers = 1
-    else:
-        max_workers = min(len(file_list), os.cpu_count() or 2)
-    errors = []
-    with ThreadPoolExecutor(max_workers=max_workers) as executor:
-        future_to_file = {
-            executor.submit(op_func, f, **{**kwargs, 'load_result': False}): f
-            for f in file_list
-        }
-        for future in as_completed(future_to_file):
-            file_path = future_to_file[future]
-            try:
-                future.result()
-            except Exception as e:
-                errors.append((file_path, e))
-
-    if errors:
-        summary = '; '.join(
-            f'{file_path} ({type(exc).__name__}: {exc})' for file_path, exc in errors
-        )
-        raise RuntimeError(f'{op_func.__name__} failed: {summary}') from errors[0][1]
-
-
-def decrypt_all_cred_files_for_env(**kwargs):
-    files = get_all_necessary_cred_files()
-    if not get_crypt():
-        check_for_encrypted_files(files)
-        return
-
-    backend = get_crypt_backend()
-    t0 = time.perf_counter()
-    _batch_cred_op(files, decrypt_file, **kwargs)
-    elapsed = time.perf_counter() - t0
-    logger.info(f'Decrypted {len(files)} cred files in {elapsed:.3f}s (backend={backend})')
-    logger.debug("Decrypted next cred files:")
-    logger.debug(files)
-
-
-def encrypt_all_cred_files_for_env(**kwargs):
-    files = get_all_necessary_cred_files()
-    logger.debug("Attempting to encrypt(if crypt is true) next files:")
-    logger.debug(files)
-    backend = get_crypt_backend()
-    t0 = time.perf_counter()
-    _batch_cred_op(files, encrypt_file, **kwargs)
-    elapsed = time.perf_counter() - t0
-    logger.info(f'Encrypted {len(files)} cred files in {elapsed:.3f}s (backend={backend})')
-
-
-def get_crypt():
-    config = get_envgene_config_yaml()
-    return config.get('crypt', True)
-
-
-def get_crypt_backend():
-    config = get_envgene_config_yaml()
-    return config.get('crypt_backend', 'Fernet')
+from envgene_shared.crypto.crypt import encrypt_all_cred_files_for_env, \
+    decrypt_all_cred_files_for_env, encrypt_file, decrypt_file, get_configured_encryption_type, \
+    is_encrypted
+from envgene_shared.utils.crypt_utils import get_crypt
+from envgene_shared.utils.file_utils import is_cred_file
